# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/branch": {
            "put": {
                "description": "Update a branch with new information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branches"
                ],
                "summary": "Update a branch",
                "parameters": [
                    {
                        "description": "Branch object to be updated",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Branch"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Branch updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Branch"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branches"
                ],
                "summary": "Create a new branch",
                "parameters": [
                    {
                        "description": "Branch object to be created",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Branch"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Branch created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Branch"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all branches",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branches"
                ],
                "summary": "Delete all branches",
                "responses": {
                    "200": {
                        "description": "All branches have been deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/branch/{id}": {
            "get": {
                "description": "Retrieve a branch by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branches"
                ],
                "summary": "Get a branch by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Branch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Branch"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a branch by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Branches"
                ],
                "summary": "Delete a branch by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Branch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Branch"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/super/bank": {
            "get": {
                "description": "Retrieve all banks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banks"
                ],
                "summary": "Get all banks",
                "responses": {
                    "200": {
                        "description": "Banks retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update a bank with new information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banks"
                ],
                "summary": "Update a bank",
                "parameters": [
                    {
                        "description": "Name of the bank",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Bank updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Bank"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new bank with the provided name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banks"
                ],
                "summary": "Create a new bank",
                "parameters": [
                    {
                        "description": "Name of the bank",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Bank created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Bank"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all banks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banks"
                ],
                "summary": "Delete all banks",
                "responses": {
                    "200": {
                        "description": "message\": \"All rows have been deleted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/super/bank/{id}": {
            "get": {
                "description": "Retrieve a bank by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banks"
                ],
                "summary": "Get a bank by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bank ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Bank"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a bank by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Banks"
                ],
                "summary": "Delete a bank by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bank ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Bank"
                        }
                    },
                    "400": {
                        "description": "error\": \"Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Account": {
            "type": "object",
            "properties": {
                "accountNumber": {
                    "type": "string"
                },
                "accountType": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "branch": {
                    "$ref": "#/definitions/models.Branch"
                },
                "branchID": {
                    "type": "integer"
                },
                "customer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Customer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "transaction": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                }
            }
        },
        "models.Bank": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Branch"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Branch": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Account"
                    }
                },
                "address": {
                    "type": "string"
                },
                "bank": {
                    "$ref": "#/definitions/models.Bank"
                },
                "bankID": {
                    "type": "integer"
                },
                "customer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Customer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "ifsc_CODE": {
                    "type": "string"
                }
            }
        },
        "models.Customer": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Account"
                    }
                },
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "branch": {
                    "$ref": "#/definitions/models.Branch"
                },
                "branchID": {
                    "type": "integer"
                },
                "dob": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pan": {
                    "type": "string"
                },
                "phone": {
                    "type": "integer"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/models.Account"
                },
                "accountID": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "modeOfPayment": {
                    "type": "string"
                },
                "receiverAccountNumber": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "typeOfTransaction": {
                    "type": "string"
                }
            }
        }
    }
}